Handle wasm plugins

+++ pkg/core/scrapper.go
@@ -299,28 +299,30 @@ func (s *Scrapper) process(ctx context.Context, repository *github.Repository) (
 		return nil, err
 	}
 
-	// Creates temp GOPATH
+	if manifest.Runtime != "wasm" {
+		// Creates temp GOPATH
 
-	gop, err := os.MkdirTemp("", "traefik-plugin-gop")
-	if err != nil {
-		span.RecordError(err)
-		return nil, fmt.Errorf("failed to create temp GOPATH: %w", err)
-	}
+		gop, err := os.MkdirTemp("", "traefik-plugin-gop")
+		if err != nil {
+			span.RecordError(err)
+			return nil, fmt.Errorf("failed to create temp GOPATH: %w", err)
+		}
 
-	defer func() { _ = os.RemoveAll(gop) }()
+		defer func() { _ = os.RemoveAll(gop) }()
 
-	// Get sources
+		// Get sources
 
-	err = s.sources.Get(ctx, repository, gop, module.Version{Path: moduleName, Version: latestVersion})
-	if err != nil {
-		span.RecordError(err)
-		return nil, fmt.Errorf("failed to get sources: %w", err)
-	}
+		err = s.sources.Get(ctx, repository, gop, module.Version{Path: moduleName, Version: latestVersion})
+		if err != nil {
+			span.RecordError(err)
+			return nil, fmt.Errorf("failed to get sources: %w", err)
+		}
 
-	// Check Yaegi interface
-	err = s.yaegiCheck(manifest, gop, moduleName)
-	if err != nil {
-		return nil, fmt.Errorf("failed to run the plugin with Yaegi: %w", err)
+		// Check Yaegi interface
+		err = s.yaegiCheck(manifest, gop, moduleName)
+		if err != nil {
+			return nil, fmt.Errorf("failed to run the plugin with Yaegi: %w", err)
+		}
 	}
 
 	snippets, err := createSnippets(repository, manifest)
@@ -332,6 +334,8 @@ func (s *Scrapper) process(ctx context.Context, repository *github.Repository) (
 	return &plugin.Plugin{
 		Name:          moduleName,
 		DisplayName:   manifest.DisplayName,
+		Runtime:       manifest.Runtime,
+		WasmPath:      manifest.WasmPath,
 		Author:        repository.GetOwner().GetLogin(),
 		RepoName:      repository.GetName(),
 		Type:          manifest.Type,
@@ -430,8 +434,12 @@ func (s *Scrapper) loadManifestContent(content string) (Manifest, error) {
 		return Manifest{}, fmt.Errorf("unsupported type: %s", m.Type)
 	}
 
-	if m.Import == "" {
+	if m.Runtime != "wasm" && m.Import == "" {
 		return Manifest{}, errors.New("missing import")
+	}
+
+	if m.Runtime == "wasm" && m.WasmPath == "" {
+		return Manifest{}, errors.New("missing wasmPath")
 	}
 
 	if m.DisplayName == "" {
@@ -889,7 +897,7 @@ func checkModuleFile(mod *modfile.File, manifest Manifest) error {
 		}
 	}
 
-	if !strings.HasPrefix(strings.ReplaceAll(manifest.Import, "-", "_"), strings.ReplaceAll(mod.Module.Mod.Path, "-", "_")) {
+	if manifest.Runtime != "wasm" && !strings.HasPrefix(strings.ReplaceAll(manifest.Import, "-", "_"), strings.ReplaceAll(mod.Module.Mod.Path, "-", "_")) {
 		return fmt.Errorf("the import %q must be related to the module name %q", manifest.Import, mod.Module.Mod.Path)
 	}
 
@@ -903,7 +911,7 @@ func checkRepoName(repository *github.Repository, moduleName string, manifest Ma
 		return fmt.Errorf("unsupported plugin: the module name (%s) doesn't contain the GitHub repository name (%s)", moduleName, repoName)
 	}
 
-	if !strings.HasPrefix(manifest.Import, repoName) {
+	if manifest.Runtime != "wasm" && !strings.HasPrefix(manifest.Import, repoName) {
 		return fmt.Errorf("unsupported plugin: the import name (%s) doesn't contain the GitHub repository name (%s)", manifest.Import, repoName)
 	}
